Review changes and misc fixes
